+++ future/utils/__init__.py
@@ -411,9 +411,27 @@ else:
 
         on Python 3. (See PEP 3134).
         """
-        exc_copy = copy.copy(exc)
-        exc_copy.__cause__ = cause
-        raise exc_copy
+        # Is either arg an exception class (e.g. IndexError) rather than
+        # instance (e.g. IndexError('my message here')? If so, pass the
+        # name of the class undisturbed through to "raise ... from ...".
+        if isinstance(exc, type) and issubclass(exc, Exception):
+            e = exc()
+            # exc = exc.__name__
+            # execstr = "e = " + _repr_strip(exc) + "()"
+            # myglobals, mylocals = _get_caller_globals_and_locals()
+            # exec(execstr, myglobals, mylocals)
+        else:
+            e = exc
+        if isinstance(cause, type) and issubclass(cause, Exception):
+            e.__cause__ = cause()
+        elif cause is None:
+            e.__cause__ = None
+        elif isinstance(cause, BaseException):
+            e.__cause__ = cause
+        else:
+            raise TypeError("exception causes must derive from BaseException")
+        e.__context__ = sys.exc_info()[1]
+        raise e
 
     exec('''
 def raise_(tp, value=None, tb=None):